+++ packages/component-library/stories/Text.stories.tsx
@@ -1,4 +1,3 @@
-import { loadElmStories } from "@cultureamp/elm-storybook"
 import { Text } from "@kaizen/component-library"
 import * as React from "react"
 
@@ -241,19 +240,3 @@ export const ZenDataSmallUnits = () => (
     %
   </Text>
 )
-
-loadElmStories("Text (Elm)", module, require("./TextStories.elm"), [
-  "h1",
-  "h2",
-  "h3",
-  "zen-display-0",
-  "zen-heading-1",
-  "zen-heading-3",
-  "zen-heading-3",
-  "zen-data-large",
-  "zen-data-large-units",
-  "zen-data-medium",
-  "zen-data-medium-units",
-  "zen-data-small",
-  "zen-data-small-units",
-])